Skip filename in Content-Disposition if it's null.
When making an upload from a byte[] there isn't a filename and the Content-Disposition header ends up with a value ending in filename="null" which can result in consuming services thinking null is should be used as a filename. Instead when it's a byte[] upload we shouldn't output the filename.

The filename part of the Content-Disposition header is optional(see RFC-6266).

Fixes #52